LayoutTests/imported/w3c/web-platform-tests/css/css-conditional/container-queries/animation-container-type-dynamic-expected.txt

Test

FAIL Animated container creating new container assert_equals: expected "rgb(0, 0, 255)" but got "rgb(255, 0, 0)"
